Move title & description from FormGenerator component up into Mutation component
