+++ src/components/graphqlFormGenerator/FormGenerator.vue
@@ -17,55 +17,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 <template>
   <v-form
-    class="c-mutation-form"
     :value="value"
     @input="$emit('input', $event)"
   >
-    <!-- the mutation title -->
-    <h3
-     style="text-transform: capitalize;"
-    >
-      {{ mutation.name }}
-    </h3>
-
-    <!-- the mutation description -->
-    <v-expansion-panels
-     accordion
-     flat
-     v-bind="extendedDescription ? { hover: true } : { readonly: true }"
-    >
-      <v-expansion-panel
-        class="mutation-desc"
-      >
-        <v-expansion-panel-header
-          v-bind="extendedDescription ? {} : {
-            expandIcon: null,
-            style: {
-              cursor: 'default'
-            }
-          }"
-        >
-          <Markdown
-           :markdown="shortDescription"
-          />
-        </v-expansion-panel-header>
-        <v-expansion-panel-content
-          v-if="extendedDescription"
-        >
-          <Markdown
-           :markdown="extendedDescription"
-          />
-        </v-expansion-panel-content>
-      </v-expansion-panel>
-    </v-expansion-panels>
-
-    <v-divider></v-divider>
-
     <!-- the form inputs -->
     <v-list>
       <v-list-item
-       v-for="input in inputs"
-       v-bind:key="input.label"
+        v-for="input in inputs"
+        v-bind:key="input.label"
       >
         <v-list-item-content>
           <v-list-item-title>
@@ -106,11 +65,7 @@ import { mdiHelpCircleOutline } from '@mdi/js'
 
 import Markdown from '@/components/Markdown'
 import FormInput from '@/components/graphqlFormGenerator/FormInput'
-import {
-  getNullValue,
-  getMutationShortDesc,
-  getMutationExtendedDesc
-} from '@/utils/aotf'
+import { getNullValue } from '@/utils/aotf'
 
 export default {
   name: 'form-generator',
@@ -124,7 +79,7 @@ export default {
     value: {
       // validity of form
       type: Boolean,
-      required: true,
+      required: false,
       default: () => false
     },
     mutation: {
@@ -167,15 +122,6 @@ export default {
         })
       }
       return ret
-    },
-
-    /* Return the first line of the description. */
-    shortDescription () {
-      return getMutationShortDesc(this.mutation.description)
-    },
-    /* Return the subsequent lines of the description */
-    extendedDescription () {
-      return getMutationExtendedDesc(this.mutation.description)
     }
   },
