+++ src/styles/cylc/_mutation_form.scss
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
- .c-mutation-form .mutation-desc {
+ .c-mutation-dialog .mutation-desc {
   .v-expansion-panel-header {
     font-size: 1em;
     line-height: inherit;
